do repackage
