+++ src/main/java/com/develhope/spring/deals/models/OrderStatus.java
@@ -1,0 +1,11 @@
+package com.develhope.spring.deals.models;
+
+/**
+ * Enum representing the various possible statuses of an order.
+ */
+public enum OrderStatus {
+    PENDING,
+    PAID,
+    SHIPPED,
+    CANCELLED
+}
