--- src/main/java/com/develhope/spring/model/OrderStatus.java
@@ -1,11 +1,0 @@
-package com.develhope.spring.model;
-
-/**
- * Enum representing the various possible statuses of an order.
- */
-public enum OrderStatus {
-    PENDING,
-    PAID,
-    SHIPPED,
-    CANCELLED
-}
