--- src/main/java/com/develhope/spring/model/vehicle/vehicleEnums/Gears.java
@@ -1,6 +1,0 @@
-package com.develhope.spring.model.vehicle.vehicleEnums;
-
-public enum Gears {
-    MANUAL,
-    AUTOMATIC
-}
