--- src/main/java/com/develhope/spring/model/vehicle/vehicleEnums/UsedFlag.java
@@ -1,6 +1,0 @@
-package com.develhope.spring.model.vehicle.vehicleEnums;
-
-public enum UsedFlag {
-    USED,
-    NEW
-}
