+++ src/main/java/com/develhope/spring/users/models/exceptions/EmptyParameterException.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.model.user;
+package com.develhope.spring.users.models.exceptions;
 
 public class EmptyParameterException extends Exception{
