+++ src/main/java/com/develhope/spring/users/models/exceptions/WrongEmailFormatException.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.model.user;
+package com.develhope.spring.users.models.exceptions;
 
 public class WrongEmailFormatException extends Exception{
