+++ src/main/java/com/develhope/spring/vehicle/vehicleEnums/Gears.java
@@ -1,0 +1,6 @@
+package com.develhope.spring.vehicle.vehicleEnums;
+
+public enum Gears {
+    MANUAL,
+    AUTOMATIC
+}
