+++ src/main/java/com/develhope/spring/vehicle/vehicleEnums/MotorPowerSupply.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.model.vehicle.vehicleEnums;
+package com.develhope.spring.vehicle.vehicleEnums;
 
 public enum MotorPowerSupply {
     DIESEL,
