+++ src/main/java/com/develhope/spring/vehicle/vehicleEnums/UsedFlag.java
@@ -1,0 +1,6 @@
+package com.develhope.spring.vehicle.vehicleEnums;
+
+public enum UsedFlag {
+    USED,
+    NEW
+}
